Two example react components to show integration between React and Astro.
With the Tooltip component, we have a react element that can be included with just a bare element (<Tooltip/>). In this case, all we have to do is install @astrojs/react + the component we want, and then make sure to add the client:only="react" to the component when we include it in our Astro file.

For the "Burger menu" we had a slightly bigger challenge -- in this case we need actual content in the React component, so we create our own react component (Burger.jsx) which does the import, and then we can use that in main. This one also required custom css, which we put in burger.css and then import in our main.js so it's present.

For the tooltip, we added data attributes throughout our project to the elements we wanted react-tooltip to effect.

For the burger menu, it's self-contained, so basically we just had to add a <Burger/> call somewhere and we get our little slidey burger menu.